Make Swagger Auth URLs https #107
