+++ tests/api/test___init__.py
@@ -81,3 +81,11 @@ def test_handle_error_unspecificRaises(app):
 
     with pytest.raises(Exception):
         api.handle_error(error)
+
+
+def test_add_oauth2_scheme(app, utils):
+    from project.api import add_oauth2_scheme_with_transport
+
+    app.config["SERVER_NAME"] = "127.0.0.1"
+    with app.app_context():
+        add_oauth2_scheme_with_transport(False)
