--- tests/test_init_data.py
@@ -1,7 +1,0 @@
-def test_add_oauth2_scheme(app, utils):
-    import os
-
-    del os.environ["AUTHLIB_INSECURE_TRANSPORT"]
-
-    url = utils.get_url("home")
-    utils.get_ok(url)
